requirements.txt
Pillow>=10.0.0
requests>=2.31.0
geocodio>=1.0.1
python-dotenv>=1.0.0
